Update add to be a replacement for write
